// types.ts
export enum RenderStyle {
  Classic = 'Classic',
  Neoclassical = 'Neoclassical',
  Modern = 'Modern',
  Minimalist = 'Minimalist',
  Brutalist = 'Brutalist',
  Industrial = 'Industrial',
  Scandinavian = 'Scandinavian',
  ArtDeco = 'Art Deco',
  Postmodern = 'Postmodern',
  Futuristic = 'Futuristic',
  Biophilic = 'Biophilic',
}

export enum TimeOfDay {
  Morning = 'Morning',
  Afternoon = 'Afternoon',
  Evening = 'Golden Hour',
  Night = 'Night',
}

export enum TimeOfYear {
  Spring = 'Spring',
  Summer = 'Summer',
  Autumn = 'Autumn',
  Winter = 'Winter',
}

export enum Weather {
  Sunny = 'Sunny',
  Cloudy = 'Cloudy',
  Rainy = 'Rainy',
  Snowy = 'Snowy',
  AfterRain = 'After Rain',
  AfterSnow = 'After Snow',
  Windy = 'Windy',
}

export enum VegetationAmount {
  None = 'None',
  Low = 'Low',
  Medium = 'Medium',
  High = 'High',
  Lush = 'Lush',
}

export enum UrbanizationLevel {
    None = 'None',
    Suburban = 'Suburban',
    Urban = 'Urban',
    Metropolis = 'Metropolis',
}

export enum PeopleDensity {
    None = 'None',
    Few = 'Few',
    Moderate = 'Moderate',
    Crowded = 'Crowded',
}

export enum VehicleDensity {
    None = 'None',
    Low = 'Low',
    Medium = 'Medium',
    High = 'High',
}

export enum PropsDensity {
    None = 'None',
    Low = 'Low',
    Medium = 'Medium',
    High = 'High',
}

export enum CameraAngle {
    AsIs = 'As Per Sketch',
    EyeLevel = 'Eye-Level View',
    HighAngle = 'High-Angle View',
    LowAngle = 'Low-Angle View',
    Front = 'Front View (Elevation)',
    Side = 'Side View (Elevation)',
    ThreeQuarter = 'Three-Quarter View',
    TopDown = 'Top-Down View',
}